Learn Node.js & Express with Project in 2 Hours p3

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require('../models/contactModel');
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler (async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 })
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler (async (req, res, next) => {
     console.log(req.body);
     const { name, email, phone } = req.body;
@@ -22,7 +22,8 @@ const createContact = asyncHandler (async (req, res, next) => {
     const contact = await Contact.create({
         name, 
         email, 
-        phone
+        phone,
+        user_id: req.user.id
     });
 
     res.status(201).json(contact);
@@ -30,13 +31,19 @@ const createContact = asyncHandler (async (req, res, next) => {
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler (async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id, 
         req.body, 
@@ -47,26 +54,38 @@ const updateContact = asyncHandler (async (req, res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler (async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts");
+    }
+
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
 })
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler (async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to view other user contacts");
+    }
+
     res.status(200).json(contact);
 })
